split repository into multiple files to handle their own table
