+++ java-itu-minitwit/src/main/java/Persistence/DB.java
@@ -1,5 +1,8 @@
 package Persistence;
 
+import Model.Follower;
+import Model.Message;
+import Model.User;
 import RoP.Failure;
 import RoP.Result;
 import RoP.Success;
@@ -9,7 +12,7 @@ public class DB {
     static final int PORT = 3306;
     private static Database instance;
     private static String database = "minitwit";
-    private static String ip = "localhost";
+    private static final String ip = "localhost";
     private static String user = "root";
     private static String pw = "root";
     private static String connectionString;
@@ -41,10 +44,6 @@ public class DB {
         return new Success<>(instance);
     }
 
-    public static void setIP(String ip) {
-        DB.ip = ip;
-    }
-
     public static void setCONNECTIONSTRING(String connectionString) {
         DB.connectionString = connectionString;
     }
@@ -59,6 +58,23 @@ public class DB {
 
     public static void setDATABASE(String dbName) {
         database = dbName;
+    }
+
+    public static Database initDb()  {
+        return DB.connectDb().get();
+    }
+    public static void dropDB(){
+        var db = initDb();
+        db.sql("drop table if exists follower").execute();
+        db.sql("drop table if exists message").execute();
+        db.sql("drop table if exists user").execute();
+
+        db.createTable(User.class);
+        db.createTable(Message.class);
+        db.sql("ALTER TABLE message ADD FOREIGN KEY (authorId) REFERENCES user(id)").execute();
+        db.createTable(Follower.class);
+        db.sql("ALTER TABLE follower ADD FOREIGN KEY (whoId) REFERENCES user(id)").execute();
+        db.sql("ALTER TABLE follower ADD FOREIGN KEY (whomId) REFERENCES user(id)").execute();
     }
 
 }
